apps/cloud-api/requirements.txt
# FastAPI + Uvicorn
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
bcrypt==4.1.2

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Redis
redis==5.0.1
celery==5.3.6

# Logging
structlog==24.1.0

# Security
cryptography==42.0.0
slowapi==0.1.9

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Utilities
python-dotenv==1.0.0

# Video streaming
opencv-python-headless==4.9.0.80
numpy<2.0
aiortc==1.9.0
av==12.3.0
aiortc==1.9.0
av
